refactoring of modules
